# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "E-commerce API Documentation",
        "version": "1.0.0",
        "description": "API for managing users, products, categories, and orders."
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "fullname": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "fullname",
                                "username",
                                "password",
                                "email"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "User already exists"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login a user",
                "parameters": [
                    {
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "username",
                                "password"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                },
                                "user": {
                                    "type": "object",
                                    "properties": {
                                        "fullname": {
                                            "type": "string"
                                        },
                                        "username": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                }
            }
        },
        "/auth/update": {
            "put": {
                "summary": "Update user information",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "fullname": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "fullname",
                                "email"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information updated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "fullname": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/category/add": {
            "post": {
                "summary": "Add a new category",
                "parameters": [
                    {
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "desc": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "name",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category added successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/category/edit/{id}": {
            "put": {
                "summary": "Edit a category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "desc": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "name",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/category/all": {
            "get": {
                "summary": "Get all categories",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/add": {
            "post": {
                "summary": "Add a new product",
                "parameters": [
                    {
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "img": {
                                    "type": "string"
                                },
                                "desc": {
                                    "type": "string"
                                },
                                "idCategory": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "required": [
                                "name",
                                "img",
                                "desc",
                                "idCategory",
                                "price"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product added successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/edit/{id}": {
            "put": {
                "summary": "Edit a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "img": {
                                    "type": "string"
                                },
                                "desc": {
                                    "type": "string"
                                },
                                "idCategory": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "required": [
                                "name",
                                "img",
                                "desc",
                                "idCategory",
                                "price"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/all": {
            "get": {
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "idCategory": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/category/{idCategory}": {
            "get": {
                "summary": "Get products by category",
                "parameters": [
                    {
                        "name": "idCategory",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products in the category",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string"
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "idCategory": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/detail/{id}": {
            "get": {
                "summary": "Get product detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "img": {
                                    "type": "string"
                                },
                                "desc": {
                                    "type": "string"
                                },
                                "idCategory": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/order/create": {
            "post": {
                "summary": "Create a new order",
                "parameters": [
                    {
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "idUser": {
                                    "type": "string"
                                },
                                "note": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "number"
                                }
                            },
                            "required": [
                                "idUser",
                                "note",
                                "total"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order created successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/order/all/status/{status}": {
            "get": {
                "summary": "Get all orders by status",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "idUser": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/order/all/user/{idUser}": {
            "get": {
                "summary": "Get all orders by user ID",
                "parameters": [
                    {
                        "name": "idUser",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "idUser": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/order/detail/{idOrder}": {
            "get": {
                "summary": "Get order detail",
                "parameters": [
                    {
                        "name": "idOrder",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "idUser": {
                                    "type": "string"
                                },
                                "note": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "number"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "orderDetails": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "idProduct": {
                                                "type": "string"
                                            },
                                            "quantity": {
                                                "type": "number"
                                            },
                                            "price": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}